requirements.txt
kucoin-python>=2.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
requests>=2.31.0
python-dateutil>=2.8.2
